Page layer prototype
